+++ changelog.md
@@ -2,6 +2,7 @@
 - Moved to new version numbering system
 - Fixed a crash with railgun & chemthrower projectiles (BluSunrize)
 - Fixed manual rendering for blueprint recipes (BluSunrize)
+- Fixed scaffolding causing a suffocation overlay (BluSunrize)
 
 ##### Version 1.15.2-3.1-pre
 - Moved to new version numbering system